+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/internal/CborSerializableStatusList.kt
@@ -7,14 +7,37 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.cbor.ByteString
 
+/**
+ * Implements [ietf-oauth-status-list](https://datatracker.ietf.org/doc/draft-ietf-oauth-status-list/) Sec. 4.3
+ */
 @Serializable
 internal data class CborSerializableStatusList(
+    /**
+     * lst: REQUIRED.  CBOR Byte string (Major Type 2) that contains
+     * the status values for all the Referenced Tokens it conveys
+     * statuses for.  The value MUST be the compressed byte array as
+     * specified in Section 4.1.
+     */
     @SerialName("lst")
     @ByteString
     val compressed: ByteArray,
-    @Serializable(with = TokenStatusBitSizeValueSerializer::class)
+
+    /**
+     * bits: REQUIRED.  CBOR Unsigned integer (Major Type 0) that
+     * contains the number of bits per Referenced Token in the
+     * compressed byte array (lst).  The allowed values for bits are
+     * 1, 2, 4 and 8.
+     */
     @SerialName("bits")
+    @Serializable(with = TokenStatusBitSizeValueSerializer::class)
     val statusBitSize: TokenStatusBitSize,
+
+    /**
+     * aggregation_uri: OPTIONAL.  CBOR Text string (Major Type 3)
+     * that contains a URI to retrieve the Status List Aggregation for
+     * this type of Referenced Token.  See section Section 9 for
+     * further detail.
+     */
     @SerialName("aggregation_uri")
     val aggregationUri: String? = null
 ) {